Refactor: Adjust padding in speed chart container
The padding within the speed chart container in `WiFiStressorScreen` has been modified.

The overall padding of the main `Padding` widget was changed from `EdgeInsets.all(16)` to `EdgeInsets.fromLTRB(1, 16, 1, 1)`.

Additionally, the `Row` containing the chart title and data point count is now wrapped in a `Container` with horizontal padding: `EdgeInsets.fromLTRB(15, 0, 15, 0)`.